Dynamic Chatbot AI/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.3.5
pydantic==2.5.0
python-multipart==0.0.6
aioredis==2.0.1
spacy==3.7.2
textblob==0.17.1
websockets==12.0